requirements.txt
python-docx==0.8.11
openpyxl==3.1.2
flask==2.3.3
werkzeug==2.3.7
